+++ src/Pages/LoginPage/LoginPage.css
@@ -19,22 +19,6 @@ h1{
     padding: 15px 20px 25px 20px;
 }
 
-.form{
-display: flex;
-align-items: center;
-justify-content: center;
-flex-direction: column;
-}
 
-.form,input{
-    margin: 0 2.5% 2.5% 2.5%;
-    border-radius: 50px;
-    border: none;
-    width: 90%;
-}
-
-input{
-    padding: 5%;
-}
