+++ src/components/Login/login.css
@@ -1,0 +1,17 @@
+.form{
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+    }
+    
+    .form,input{
+        margin: 0 2.5% 2.5% 2.5%;
+        border-radius: 50px;
+        border: none;
+        width: 90%;
+    }
+    
+    input{
+        padding: 5%;
+    }
